unify pasted/scanned content.
